修改胡牌判断方法，判断正式番、组合番时，如果存在还得排除互斥的番型 修复清一色判断问题 规则中加入 特殊番型集合（直杠、补杠、暗杠、退税、查叫、查花猪 的番型id）。后面可以新增一个SichuanRule，挪到里面去

+++ common/src/main/java/com/yude/game/common/model/history/SettlementInfo.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class SettlementInfo {
     private Integer posId;
     private Long beforeScore;
-    private Long changeScore;
+    private int changeScore;
     private Long remaningScore;
     private int fanNum;
     private List<Integer> standCards;
@@ -40,11 +40,11 @@ public class SettlementInfo {
         return this;
     }
 
-    public Long getChangeScore() {
+    public int getChangeScore() {
         return changeScore;
     }
 
-    public SettlementInfo setChangeScore(Long changeScore) {
+    public SettlementInfo setChangeScore(int changeScore) {
         this.changeScore = changeScore;
         return this;
     }
